+++ src/com/bsb/hike/utils/HikeToast.java
@@ -23,7 +23,7 @@ public class HikeToast {
         this.notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
     }
 
-    public void toast(ContactInfo contactInfo, String msisdn, String message, int timestamp) {
+    public void toast(ContactInfo contactInfo, String msisdn, String message, long timestamp) {
         String name = (contactInfo != null) ? contactInfo.name : msisdn;
         int icon = R.drawable.ic_launcher;
